+++ cli/src/cluster_query.rs
@@ -1446,6 +1446,7 @@ pub async fn process_show_block_production(
     Ok(config.output_format.formatted_string(&block_production))
 }
 
+#[allow(deprecated)]
 pub async fn process_largest_accounts(
     rpc_client: &RpcClient,
     config: &CliConfig<'_>,