+++ rpc-client/src/rpc_client.rs
@@ -1846,6 +1846,7 @@ impl RpcClient {
     /// # Examples
     ///
     /// ```
+    /// # #![allow(deprecated)]
     /// # use solana_rpc_client_api::{
     /// #     client_error::Error,
     /// #     config::{RpcLargestAccountsConfig, RpcLargestAccountsFilter},
@@ -1864,6 +1865,12 @@ impl RpcClient {
     /// )?;
     /// # Ok::<(), Error>(())
     /// ```
+    #[allow(deprecated)]
+    #[deprecated(
+        since = "4.1.0",
+        note = "This uses the deprecated `getLargestAccounts` RPC method and may be removed in a \
+                future major release"
+    )]
     pub fn get_largest_accounts_with_config(
         &self,
         config: RpcLargestAccountsConfig,